+++ src/app/layout.tsx
@@ -10,6 +10,7 @@ import { bgSubPrimary } from '@/styles/ogoo/colors.css';
 import { mainContainer } from '@/styles/ogoo/container.css';
 import { cn } from '@/utils';
 
+import Kakaoinit from './kakaoinit';
 import { NextAuthProvider } from './providers';
 
 const Pretendard = localFont({
@@ -20,6 +21,12 @@ export const metadata: Metadata = {
   title: '5959',
   description: '오구오구에 오신 것을 환영합니다',
 };
+
+declare global {
+  interface Window {
+    Kakao: any;
+  }
+}
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
@@ -53,20 +60,15 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
             `,
           }}
         />
-        {/* <Script
-          defer
-          src="https://t1.kakaocdn.net/kakao_js_sdk/2.4.0/kakao.min.js"
-          integrity="sha384-mXVrIX2T/Kszp6Z0aEWaA8Nm7J6/ZeWXbL8UpGRjKwWe56Srd/iyNmWMBhcItAjH"
-          crossOrigin="anonymous"
-        ></Script> */}
-        <Script defer src="https://developers.kakao.com/sdk/js/kakao.min.js" />
       </head>
 
       <body className={cn(themeClass, Pretendard.className, flexCenter, bgSubPrimary)}>
         <main
           className={cn(mainContainer, `flex-1 max-w-md relative overflow-hidden drop-shadow-sm`)}
         >
-          <NextAuthProvider>{children}</NextAuthProvider>
+          <Kakaoinit>
+            <NextAuthProvider>{children}</NextAuthProvider>
+          </Kakaoinit>
         </main>
       </body>
     </html>
